move addr label to common
