+++ apps/daimo-mobile/package.json
@@ -76,7 +76,7 @@
     "expo-image": "^1.10.6",
     "expo-linking": "~6.2.2",
     "expo-local-authentication": "~13.8.0",
-    "expo-localization": "~14.8.4",
+    "expo-localization": "^14.8.4",
     "expo-notifications": "~0.27.6",
     "expo-splash-screen": "~0.26.4",
     "expo-status-bar": "~1.11.1",
