+++ apps/daimo-mobile/src/i18n/languages/en.ts
@@ -1,5 +1,3 @@
-import { AddrLabel } from "@daimo/common";
-
 export const en = {
   // Common text components
   shared: {
@@ -768,11 +766,6 @@ export const en = {
   },
 
   // -------------------------- MISC --------------------------
-  // AddrLabels for account history contacts
-  addrLabel: {
-    label: (type: AddrLabel) => displayAddrLabel(type),
-  },
-
   tabNav: {
     deposit: () => `Deposit`,
     invite: () => `Invite`,
@@ -785,32 +778,4 @@ export const en = {
 function pluralize(n: number, noun: string) {
   if (n === 1) return `${n} ${noun}`; // "1 apple"
   return `${n} ${noun}s`; // "0 apples" or "2 apples"
-}
-
-// Addr label display function
-function displayAddrLabel(type: AddrLabel): string {
-  switch (type) {
-    case AddrLabel.Faucet:
-      return `team Daimo`;
-    case AddrLabel.PaymentLink:
-      return `payment link`;
-    case AddrLabel.RequestLink:
-      return `request link`;
-    case AddrLabel.Paymaster:
-      return `fee`;
-    case AddrLabel.Coinbase:
-      return `coinbase`;
-    case AddrLabel.Relay:
-      return `relay.link`;
-    case AddrLabel.LiFi:
-      return `li.fi bridge`;
-    case AddrLabel.UniswapETHPool:
-      return `swapped ETH`;
-    case AddrLabel.Binance:
-      return `binance`;
-    case AddrLabel.FastCCTP:
-      return `instant cross-chain`;
-    default:
-      throw new Error(`Invalid AddrLabel: ${type}`);
-  }
 }
