+++ apps/daimo-mobile/src/view/screen/link/NoteScreen.tsx
@@ -26,7 +26,7 @@ import {
   useSendWithDeviceKeyAsync,
 } from "../../../action/useSendAsync";
 import { ParamListHome, useExitBack } from "../../../common/nav";
-import { i18n } from "../../../i18n";
+import { getI18NLocale, i18n } from "../../../i18n";
 import { useFetchLinkStatus } from "../../../logic/linkStatus";
 import { useEphemeralSignature } from "../../../logic/note";
 import { getRpcFunc } from "../../../logic/trpc";
@@ -119,11 +119,12 @@ function NoteDisplayInner({
   hideAmount?: boolean;
   leaveScreen?: () => void;
 }) {
+  const locale = getI18NLocale();
   // Where the note came from
   const sendPhrase =
     noteStatus.sender.addr === account.address
       ? i18.send.self()
-      : i18.send.other(getAccountName(noteStatus.sender));
+      : i18.send.other(getAccountName(noteStatus.sender, locale));
 
   // The note itself and signature
   const ephemeralOwner = noteStatus.ephemeralOwner!;
@@ -235,7 +236,7 @@ function NoteDisplayInner({
       case DaimoNoteState.Claimed:
         return (
           <TextBold>
-            {i18.accepted.long(getAccountName(noteStatus.claimer!))}
+            {i18.accepted.long(getAccountName(noteStatus.claimer!, locale))}
           </TextBold>
         );
       case DaimoNoteState.Cancelled:
